added a function that allows users to click on frame
#### Step 2: Add a second click to refine the prediction

Hmm, it seems that although we wanted to segment the child on the left, the model predicts the mask for only the shorts -- this can happen since there is ambiguity from a single click about what the target object should be. We can refine the mask on this frame via another positive click on the child's shirt.

Here we make a **second positive click** at (x, y) = (250, 220) with label `1` to expand the mask.

Note: we need to send **all the clicks and their labels** (i.e. not just the last click) when calling `add_new_points_or_box`.
Co-Authored-By: Lucas Caceres <105500229+Mxmeow@users.noreply.github.com>
Co-Authored-By: Jeanmarck12 <144716961+Jeanmarck12@users.noreply.github.com>