Updated code
Lucas fixed the zooming problem, and i created a label folder that stored the coordinates of the bounding boxes in a format that allows yolo to understand it.

Tanesha I need you to test it and see if it works. Once it completes making a label folder i want you to make a function that will go through the coordinates in the files and put them on the images to see if it actually stored the correct coordinates.